Revert "Decimal"
This reverts commit c8c3eb8f5bc2c231a431f5d765f845c5f4334442.

+++ models/character.py
@@ -48,7 +48,7 @@ class Attribute(models.Model):
 	name = models.CharField(max_length=50)
 	owner = models.ForeignKey(Character)
 	level = models.IntegerField()
-	exp = models.DecimalField(max_digits=15, decimal_places=2)
+	exp = models.FloatField()
 	class Meta:
 		app_label = 'kovin'
 	def __unicode__(self):
